address Simon's feedback

+++ lib/test/ClientTest.hs
@@ -6,7 +6,6 @@ import Control.Monad
 import Control.Monad.Trans.Class
 import Control.Monad.Trans.Reader
 import Data.IORef
-import Data.Int
 import Data.Proxy
 import TestRunner
 import Test.HUnit hiding (State)
@@ -97,7 +96,7 @@ runCalculatorServer proxy ch = do
   runServer proxy ch $ processCommand state
 
 -- Server Implementation
-type State = IORef Int64
+type State = IORef Int
 
 initServerState :: IO State
 initServerState = newIORef 0
